+++ udata_front/tests/views/test_dataset.py
@@ -74,8 +74,6 @@ class DatasetBlueprintTest(GouvfrFrontTestCase):
                          resource.created_at.isoformat()[:16])
         self.assertEqual(json_ld_resource['dateModified'][:16],
                          resource.modified.isoformat()[:16])
-        self.assertEqual(json_ld_resource['datePublished'][:16],
-                         resource.published.isoformat()[:16])
         self.assertEqual(json_ld_resource['encodingFormat'], 'png')
         self.assertEqual(json_ld_resource['contentSize'],
                          resource.filesize)
@@ -103,8 +101,6 @@ class DatasetBlueprintTest(GouvfrFrontTestCase):
                          community_resource.created_at.isoformat()[:16])
         self.assertEqual(json_ld_resource['dateModified'][:16],
                          community_resource.modified.isoformat()[:16])
-        self.assertEqual(json_ld_resource['datePublished'][:16],
-                         community_resource.published.isoformat()[:16])
         self.assertEqual(json_ld_resource['encodingFormat'],
                          community_resource.format)
         self.assertEqual(json_ld_resource['contentSize'],